Finished map showing players locations in real(ish) time. Started figuring out how to load items that an individual has picked up
Note!
Potentially creating a local DB of the items for the current player will help when choosing to load items into a scene based on the player. When player logs in, pull from online DB the items using the players ID. Then when loading each location call from the local DB using PlayerID and LocationID